[MULI-2] - Create User Account [LJ,MM]

+++ MulatschakTracker/app/src/main/java/com/example/mulatschaktracker/UserObject.kt
@@ -4,6 +4,6 @@ import android.content.Context
 import android.database.sqlite.SQLiteOpenHelper
 
 class UserObject (val name: String) {
-    var id: Int = -1
+    var id: Long = -1
 
 }
